% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csmIndexer.R
\name{csmIndexer}
\alias{csmIndexer}
\title{CSM Indexer}
\usage{
csmIndexer()
}
\arguments{
\item{path}{The path to the folder to be indexed}

\item{pageLogo}{The path to logos to add at the top of each page}

\item{titleLogo}{The path to logos to add on the title page}

\item{author}{The author of the document}

\item{title}{The document title}

\item{personnel}{The key personnel working on the documents contained in the folder. You may wish to include start and end dates here if appropriate.}
}
\description{
A small indexer which produces a pdf of the contents of a folder.
}
\details{
Prerequesites: You must have installed miktex and it must be available on the PATH.

Please set the working directory explicitely using setwd() to ensure you know where the file is being created.

Note that the R programming language uses bachslash \ to create special characters. The simplest way to write paths is to replace them with forwardslash /.
}
\examples{
#################################################################
# Note that the example has been designed to run on the I: drive
# at CSM and so will not work if you do not have access to this drive.
#################################################################

# install package containing the indexer.
devtools::install_github("csmoxford/csmIndexer")

# Load the library containing the functions
library(csmIndexer)

#################################################################
# Note that the R programming language uses bachslash \\ to create special
# characters. The simplest way to write paths is to replace them with
# forwardslash /.
#################################################################

# Tell R where to create the pdf file.
setwd("I:/Data/Useful Programs/Statistical TMF Index")

# The path to the folder to be indexed.
path <- "I:/Data/Useful Programs/Statistical TMF Index"
# The path to logos to add at the top of each page. Multi logos possible.
pageLogo <- c("I:/Data/OCTRU_SOPs/CSMLogo - Copy.png")
# The path to logos to add on the title page. Multi logos possible.
titleLogo <- c("I:/Data/OCTRU_SOPs/CSMLogo - Copy.png")
# The author of the document.
author <- "Eleni Franou"
# The document title.
title <- "Index of useful Programs Folder"
# The key personnel working on the documents contained in the folder.
# You may wish to include start and end dates here if appropriate.
personnel <- c("Peter Dutton (Statistician)", "Eleni Frangou (Statistician)")

# Create the pdf
csmIndexer()

}
\author{
Eleni Frangou

Peter Dutton
}

